+++ pkg/storage/gcs.go
@@ -20,7 +20,7 @@ import (
 	"google.golang.org/api/option/internaloption"
 
 	"cloud.google.com/go/storage"
-	"github.com/apex/log"
+	apexLog "github.com/apex/log"
 	"google.golang.org/api/option"
 	googleHTTPTransport "google.golang.org/api/transport/http"
 )
@@ -47,11 +47,11 @@ func (w debugGCSTransport) RoundTrip(r *http.Request) (*http.Response, error) {
 			logMsg += fmt.Sprintf("%v: %v\n", h, v)
 		}
 	}
-	log.Info(logMsg)
+	apexLog.Info(logMsg)
 
 	resp, err := w.base.RoundTrip(r)
 	if err != nil {
-		log.Errorf("GCS_ERROR: %v", err)
+		apexLog.Errorf("GCS_ERROR: %v", err)
 		return resp, err
 	}
 	logMsg = fmt.Sprintf("<<< [GCS_RESPONSE] <<< %v %v\n", r.Method, r.URL.String())
@@ -60,7 +60,7 @@ func (w debugGCSTransport) RoundTrip(r *http.Request) (*http.Response, error) {
 			logMsg += fmt.Sprintf("%v: %v\n", h, v)
 		}
 	}
-	log.Info(logMsg)
+	apexLog.Info(logMsg)
 	return resp, err
 }
 
@@ -225,7 +225,7 @@ func (gcs *GCS) WalkAbsolute(ctx context.Context, rootPath string, recursive boo
 func (gcs *GCS) GetFileReader(ctx context.Context, key string) (io.ReadCloser, error) {
 	pClientObj, err := gcs.clientPool.BorrowObject(ctx)
 	if err != nil {
-		log.Errorf("gcs.GetFileReader: gcs.clientPool.BorrowObject error: %+v", err)
+		apexLog.Errorf("gcs.GetFileReader: gcs.clientPool.BorrowObject error: %+v", err)
 		return nil, err
 	}
 	pClient := pClientObj.(*clientObject).Client
@@ -233,12 +233,12 @@ func (gcs *GCS) GetFileReader(ctx context.Context, key string) (io.ReadCloser, e
 	reader, err := obj.NewReader(ctx)
 	if err != nil {
 		if pErr := gcs.clientPool.InvalidateObject(ctx, pClientObj); pErr != nil {
-			log.Warnf("gcs.GetFileReader: gcs.clientPool.InvalidateObject error: %v ", pErr)
+			apexLog.Warnf("gcs.GetFileReader: gcs.clientPool.InvalidateObject error: %v ", pErr)
 		}
 		return nil, err
 	}
 	if pErr := gcs.clientPool.ReturnObject(ctx, pClientObj); pErr != nil {
-		log.Warnf("gcs.GetFileReader: gcs.clientPool.ReturnObject error: %v ", pErr)
+		apexLog.Warnf("gcs.GetFileReader: gcs.clientPool.ReturnObject error: %v ", pErr)
 	}
 	return reader, nil
 }
@@ -250,7 +250,7 @@ func (gcs *GCS) GetFileReaderWithLocalPath(ctx context.Context, key, _ string) (
 func (gcs *GCS) PutFile(ctx context.Context, key string, r io.ReadCloser) error {
 	pClientObj, err := gcs.clientPool.BorrowObject(ctx)
 	if err != nil {
-		log.Errorf("gcs.PutFile: gcs.clientPool.BorrowObject error: %+v", err)
+		apexLog.Errorf("gcs.PutFile: gcs.clientPool.BorrowObject error: %+v", err)
 		return err
 	}
 	pClient := pClientObj.(*clientObject).Client
@@ -264,17 +264,17 @@ func (gcs *GCS) PutFile(ctx context.Context, key string, r io.ReadCloser) error 
 	}
 	defer func() {
 		if err := gcs.clientPool.ReturnObject(ctx, pClientObj); err != nil {
-			log.Warnf("gcs.PutFile: gcs.clientPool.ReturnObject error: %+v", err)
+			apexLog.Warnf("gcs.PutFile: gcs.clientPool.ReturnObject error: %+v", err)
 		}
 	}()
 	buffer := make([]byte, 128*1024)
 	_, err = io.CopyBuffer(writer, r, buffer)
 	if err != nil {
-		log.Warnf("gcs.PutFile: can't copy buffer: %+v", err)
+		apexLog.Warnf("gcs.PutFile: can't copy buffer: %+v", err)
 		return err
 	}
 	if err = writer.Close(); err != nil {
-		log.Warnf("gcs.PutFile: can't close writer: %+v", err)
+		apexLog.Warnf("gcs.PutFile: can't close writer: %+v", err)
 		return err
 	}
 	return nil
@@ -298,7 +298,7 @@ func (gcs *GCS) StatFile(ctx context.Context, key string) (RemoteFile, error) {
 func (gcs *GCS) deleteKey(ctx context.Context, key string) error {
 	pClientObj, err := gcs.clientPool.BorrowObject(ctx)
 	if err != nil {
-		log.Errorf("gcs.deleteKey: gcs.clientPool.BorrowObject error: %+v", err)
+		apexLog.Errorf("gcs.deleteKey: gcs.clientPool.BorrowObject error: %+v", err)
 		return err
 	}
 	pClient := pClientObj.(*clientObject).Client
@@ -306,12 +306,12 @@ func (gcs *GCS) deleteKey(ctx context.Context, key string) error {
 	err = object.Delete(ctx)
 	if err != nil {
 		if pErr := gcs.clientPool.InvalidateObject(ctx, pClientObj); pErr != nil {
-			log.Warnf("gcs.deleteKey: gcs.clientPool.InvalidateObject error: %+v", pErr)
+			apexLog.Warnf("gcs.deleteKey: gcs.clientPool.InvalidateObject error: %+v", pErr)
 		}
 		return err
 	}
 	if pErr := gcs.clientPool.ReturnObject(ctx, pClientObj); pErr != nil {
-		log.Warnf("gcs.deleteKey: gcs.clientPool.ReturnObject error: %+v", pErr)
+		apexLog.Warnf("gcs.deleteKey: gcs.clientPool.ReturnObject error: %+v", pErr)
 	}
 	return nil
 }
@@ -328,10 +328,10 @@ func (gcs *GCS) DeleteFileFromObjectDiskBackup(ctx context.Context, key string) 
 
 func (gcs *GCS) CopyObject(ctx context.Context, srcSize int64, srcBucket, srcKey, dstKey string) (int64, error) {
 	dstKey = path.Join(gcs.Config.ObjectDiskPath, dstKey)
-	log.Debugf("GCS->CopyObject %s/%s -> %s/%s", srcBucket, srcKey, gcs.Config.Bucket, dstKey)
+	apexLog.Debugf("GCS->CopyObject %s/%s -> %s/%s", srcBucket, srcKey, gcs.Config.Bucket, dstKey)
 	pClientObj, err := gcs.clientPool.BorrowObject(ctx)
 	if err != nil {
-		log.Errorf("gcs.CopyObject: gcs.clientPool.BorrowObject error: %+v", err)
+		apexLog.Errorf("gcs.CopyObject: gcs.clientPool.BorrowObject error: %+v", err)
 		return 0, err
 	}
 	pClient := pClientObj.(*clientObject).Client
@@ -340,18 +340,18 @@ func (gcs *GCS) CopyObject(ctx context.Context, srcSize int64, srcBucket, srcKey
 	attrs, err := src.Attrs(ctx)
 	if err != nil {
 		if pErr := gcs.clientPool.InvalidateObject(ctx, pClientObj); pErr != nil {
-			log.Warnf("gcs.CopyObject: gcs.clientPool.InvalidateObject error: %+v", pErr)
+			apexLog.Warnf("gcs.CopyObject: gcs.clientPool.InvalidateObject error: %+v", pErr)
 		}
 		return 0, err
 	}
 	if _, err = dst.CopierFrom(src).Run(ctx); err != nil {
 		if pErr := gcs.clientPool.InvalidateObject(ctx, pClientObj); pErr != nil {
-			log.Warnf("gcs.CopyObject: gcs.clientPool.InvalidateObject error: %+v", pErr)
+			apexLog.Warnf("gcs.CopyObject: gcs.clientPool.InvalidateObject error: %+v", pErr)
 		}
 		return 0, err
 	}
 	if pErr := gcs.clientPool.ReturnObject(ctx, pClientObj); pErr != nil {
-		log.Warnf("gcs.CopyObject: gcs.clientPool.ReturnObject error: %+v", pErr)
+		apexLog.Warnf("gcs.CopyObject: gcs.clientPool.ReturnObject error: %+v", pErr)
 	}
 	return attrs.Size, nil
 }